+++ src/test/java/orderhardcode/OrderNegativeParamTest.java
@@ -1,4 +1,4 @@
-package order;
+package orderhardcode;
 
 import client.OrderClient;
 import client.UserClient;
@@ -6,6 +6,7 @@ import io.qameta.allure.Description;
 import io.qameta.allure.Step;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
+import order.NewOrder;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
